+++ examples/2_hw_async_actions.py
@@ -3,7 +3,7 @@ import time
 import numpy as np
 
 from home_robot.motion.stretch import STRETCH_HOME_Q, HelloStretchKinematics
-from home_robot_hw.stretch_client import StretchClient
+from home_robot_hw.remote import StretchClient
 
 if __name__ == "__main__":
     robot = StretchClient()
